Rename loader to configLoader, optimize the object creation when loading from Firebase, rename an environment variable
